+++ package.json
@@ -16,6 +16,8 @@
   },
   "dependencies": {
     "@astrojs/check": "^0.9.4",
+    "@astrojs/node": "^9.0.1",
+    "@astrojs/react": "^4.1.5",
     "@astrojs/rss": "^4.0.9",
     "@astrojs/sitemap": "^3.2.1",
     "@astrojs/svelte": "^6.0.1",
@@ -28,10 +30,13 @@
     "@iconify-json/fa6-solid": "^1.2.1",
     "@iconify-json/material-symbols": "^1.2.5",
     "@iconify/svelte": "^4.0.2",
+    "@mdxeditor/editor": "^3.20.0",
     "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@swup/astro": "^1.4.1",
     "@tailwindcss/typography": "^0.5.16",
     "@types/hast": "^3.0.4",
+    "@types/react": "^19.0.7",
+    "@types/react-dom": "^19.0.3",
     "@types/unist": "^3.0.3",
     "@upstash/redis": "^1.34.3",
     "@vcarl/remark-headings": "^0.1.0",
@@ -48,6 +53,8 @@
     "overlayscrollbars": "^2.10.0",
     "pagefind": "^1.1.1",
     "photoswipe": "^5.4.4",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "reading-time": "^1.5.0",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-components": "^0.3.0",
