+++ src/components/control/Pagination.astro
@@ -48,6 +48,7 @@ const getPageUrl = (p: number) => {
   if (p == 1) return '/'
   return `/${p}/`
 }
+
 ---
 
 <div class:list={[className, "flex flex-row gap-3 justify-center"]} style={style}>
