+++ src/components/widget/Tags.astro
@@ -5,9 +5,13 @@ import ButtonTag from '../control/ButtonTag.astro'
 import { i18n } from '../../i18n/translation'
 import I18nKey from '../../i18n/i18nKey'
 import { url } from '../../utils/url-utils'
-import { getAllTags } from '@utils/content-utils'
+import { safeFetch } from '@utils/content-utils'
 
-const tags = await getAllTags()
+const response = await safeFetch(
+  `${import.meta.env.NEXT_PUBLIC_URL}/api/posts/fetchAllTags`
+);
+
+const tags = response.result;
 
 const COLLAPSED_HEIGHT = '7.5rem'
 
@@ -22,7 +26,7 @@ const style = Astro.props.style
 ---
 <WidgetLayout name={i18n(I18nKey.tags)} id="tags" isCollapsed={isCollapsed} collapsedHeight={COLLAPSED_HEIGHT} class={className} style={style}>
     <div class="flex gap-2 flex-wrap">
-        {tags.map(t => (
+        {tags.map((t: string) => (
             <ButtonTag href={url(`/archive/tag/${t}/`)} label={`View all posts with the ${t} tag`}>
                 {t}
             </ButtonTag>
