+++ src/layouts/MainGridLayout.astro
@@ -9,7 +9,7 @@ import { siteConfig } from '../config'
 import TOC from "../components/widget/TOC.astro";
 import ImageWrapper from "../components/misc/ImageWrapper.astro";
 import {BANNER_HEIGHT, BANNER_HEIGHT_EXTEND, MAIN_PANEL_OVERLAPS_BANNER_HEIGHT} from "../constants/constants";
-import type { Headings } from '@utils/content-utils'
+import type { Headings } from '@/types/posts'
 
 interface Props {
   title?: string
