+++ src/pages/[...page].astro
@@ -4,21 +4,29 @@ import PostPage from '../components/PostPage.astro'
 import Pagination from '../components/control/Pagination.astro'
 import { PAGE_SIZE } from '../constants/constants'
 import MainGridLayout from '../layouts/MainGridLayout.astro'
-import { getSortedPosts } from '../utils/content-utils'
+import { getSortedPosts, safeFetch } from '../utils/content-utils'
 import Content from 'node_modules/photoswipe/dist/types/slide/content'
 
-export const getStaticPaths = (async ({ paginate }) => {
-  const allBlogPosts = await getSortedPosts()
-  return paginate(allBlogPosts, { pageSize: PAGE_SIZE })
-}) satisfies GetStaticPaths
-// https://github.com/withastro/astro/issues/6507#issuecomment-1489916992
+// export const prerender = true;
 
-const { page } = Astro.props
+// export const getStaticPaths = (async ({ paginate }) => {
+//   const allBlogPosts = await getSortedPosts()
+//   return paginate(allBlogPosts, { pageSize: PAGE_SIZE })
+// }) satisfies GetStaticPaths
+// // https://github.com/withastro/astro/issues/6507#issuecomment-1489916992
 
-const len = page.data.length
+//const { page } = Astro.props
+
+const response = await safeFetch(`${import.meta.env.NEXT_PUBLIC_URL}/api/posts/fetchAllPosts`)
+
+let postList = response.result
+
+const len = postList.length
+
 ---
 
 <MainGridLayout>
-    <PostPage page={page}></PostPage>
-    <Pagination class="mx-auto onload-animation" page={page} style=`animation-delay: calc(var(--content-delay) + ${(len)*50}ms)`></Pagination>
+    <PostPage page={postList}></PostPage>
+    <!-- I don't know how this Pagination component is meant to work soooooooo... Commented! -->
+    <!-- <Pagination class="mx-auto onload-animation" page={postList} style=`animation-delay: calc(var(--content-delay) + ${(len)*50}ms)`></Pagination> -->
 </MainGridLayout>
