+++ src/utils/content-utils.ts
@@ -1,104 +1,5 @@
-import I18nKey from "@i18n/i18nKey";
-import { i18n } from "@i18n/translation";
-import type { VFile } from "node_modules/rehype-raw/lib";
-import { parse } from "./parser";
 import { public_handle } from "@/config";
-
-export interface Profile {
-  avatar: string;
-  banner: string;
-  displayName: string;
-  did: string;
-  handle: string;
-  description: string;
-  pds: string;
-}
-
-export interface Ogp {
-  url: string;
-  width?: number;
-  height?: number;
-  [k: string]: unknown;
-}
-
-export interface Post {
-  title?: string;
-  rkey?: string;
-  createdAt?: Date;
-  content?: string | VFile; // content parsed to html
-  visibility?: boolean;
-  ogp?: Ogp;
-  extendedData?: PostExtended;
-  nextPost?: PostRef;
-  prevPost?: PostRef;
-  readingTime?: string;
-}
-
-export interface MarkdownPost {
-  title: string;
-  rkey: string;
-  createdAt: Date;
-  mdcontent: string; // markdown content
-  visibility: string;
-  ogp: Ogp;
-  data: any;
-}
-
-export interface ReadingTime {
-  text: number;
-  minutes: number;
-  time: number;
-  words: number;
-}
-
-export interface Headings {
-  depth: number;
-  value: string;
-}
-
-export interface PostExtended {
-  title?: string;
-  published?: string;
-  updated?: Date;
-  description?: string;
-  image?: string;
-  tags?: string[];
-  category?: string;
-  lang?: string;
-  readingTime?: ReadingTime;
-  headings?: Headings[];
-  nextSlug?: string;
-  nextTitle?: string;
-  prevSlug?: string;
-  prevTitle?: string;
-}
-
-interface PostRef {
-  title?: string;
-  slug?: string;
-}
-
-export interface PostList {
-  slug?: string;
-  body?: string | VFile;
-  data?: PostExtended;
-  lastUpdate?: Date;
-}
-
-export async function getAllTags() {
-  const response = await safeFetch(`${import.meta.env.NEXT_PUBLIC_URL}/api/posts/fetchAllPosts`);
-
-  let postList = response.result;
-  let tags: string[] = new Array();
-
-  for (const rkey in postList) {
-    for (let tag of postList[rkey].extendedData?.tags ?? []) {
-      tags.push(tag);
-    }
-  }
-
-  return tags;
-}
+import type { PostList } from "@/types/posts";
 
 export async function getSortedPosts() {
   const response = await safeFetch(`${import.meta.env.NEXT_PUBLIC_URL}/api/posts/fetchAllPosts`);
@@ -131,41 +32,6 @@ export async function getSortedPosts() {
   }
 
   return posts;
-}
-
-export type Category = {
-  name: string;
-  count: number;
-};
-
-export async function getCategoryList(): Promise<Category[]> {
-  const count: { [key: string]: number } = {};
-  const response = await safeFetch(`${import.meta.env.NEXT_PUBLIC_URL}/api/posts/fetchAllPosts`);
-
-  let postList = response.result;
-
-  for (const rkey in postList) {
-    if (!postList[rkey].extendedData.category) {
-      const ucKey = i18n(I18nKey.uncategorized);
-      count[ucKey] = count[ucKey] ? count[ucKey] + 1 : 1;
-    } else {
-      count[postList[rkey].extendedData?.category] = count[
-        postList[rkey].extendedData?.category
-      ]
-        ? count[postList[rkey].extendedData?.category] + 1
-        : 1;
-    }
-  }
-
-  const lst = Object.keys(count).sort((a, b) => {
-    return a.toLowerCase().localeCompare(b.toLowerCase());
-  });
-
-  const ret: Category[] = [];
-  for (const c of lst) {
-    ret.push({ name: c, count: count[c] });
-  }
-  return ret;
 }
 
 export async function safeFetch(url: string) {
